test: introduce test-services workflow
Introduces a new test-services workflow.

The workflow runs on push to main and pull requests that modify docker-compose.yml, balena.yml, or any file in the services directory.

The workflow runs Docker Compose with the balena-interface and wifi services disabled, waits for the services to start, takes screenshots of the services, and uploads the screenshots to S3.